Implemented Spring JPA

+++ jk-spring-jpa-poc/src/main/java/jk/research/spring/jpa/common/cache/service/CacheRefRegService.java
@@ -10,7 +10,7 @@ public interface CacheRefRegService {
 
 	Optional<CacheRefreshRegistration> findAllCacheRegistryById(CacheRefreshRegistrationId refRegId);
 
-	long deleteByIdTargetApplication(TargetApplication targetApplication);
+	Integer deleteByIdTargetApplication(TargetApplication targetApplication);
 
 	void save(CacheRefreshRegistration cacheReg);
